Made the app docker friendly Updated the docker-compose.yml file Updated the init.py file to use the docker-compose.yml file Removed UIComponents.py Removed functions.py Removed keys.py Removed test.py
